+++ packages/vite/src/node/plugins/importAnalysis.ts
@@ -254,9 +254,9 @@ export function importAnalysisPlugin(config: ResolvedConfig): Plugin {
       // since we are already in the transform phase of the importer, it must
       // have been loaded so its entry is guaranteed in the module graph.
       const importerModule = moduleGraph.getModuleById(importer)!
-      if (!importerModule && depsOptimizer?.isOptimizedDepFile(importer)) {
-        // Ids of optimized deps could be invalidated and removed from the graph
-        // Return without transforming, this request is no longer valid, a full reload
+      if (!importerModule) {
+        // When the server is restarted, the module graph is cleared, so we
+        // return without transforming. This request is no longer valid, a full reload
         // is going to request this id again. Throwing an outdated error so we
         // properly finish the request with a 504 sent to the browser.
         throwOutdatedRequest(importer)