+++ packages/vite/src/node/ssr/ssrStacktrace.ts
@@ -1,30 +1,39 @@
+import path from 'node:path'
 import { TraceMap, originalPositionFor } from '@jridgewell/trace-mapping'
 import type { ModuleGraph } from '../server/moduleGraph'
 
 let offset: number
-try {
-  new Function('throw new Error(1)')()
-} catch (e) {
-  // in Node 12, stack traces account for the function wrapper.
-  // in Node 13 and later, the function wrapper adds two lines,
-  // which must be subtracted to generate a valid mapping
-  const match = /:(\d+):\d+\)$/.exec(e.stack.split('\n')[1])
-  offset = match ? +match[1] - 1 : 0
+
+function calculateOffsetOnce() {
+  if (offset !== undefined) {
+    return
+  }
+
+  try {
+    new Function('throw new Error(1)')()
+  } catch (e) {
+    // in Node 12, stack traces account for the function wrapper.
+    // in Node 13 and later, the function wrapper adds two lines,
+    // which must be subtracted to generate a valid mapping
+    const match = /:(\d+):\d+\)$/.exec(e.stack.split('\n')[1])
+    offset = match ? +match[1] - 1 : 0
+  }
 }
 
 export function ssrRewriteStacktrace(
   stack: string,
   moduleGraph: ModuleGraph,
 ): string {
+  calculateOffsetOnce()
   return stack
     .split('\n')
     .map((line) => {
       return line.replace(
         /^ {4}at (?:(\S.*?)\s\()?(.+?):(\d+)(?::(\d+))?\)?/,
-        (input, varName, url, line, column) => {
-          if (!url) return input
+        (input, varName, id, line, column) => {
+          if (!id) return input
 
-          const mod = moduleGraph.urlToModuleMap.get(url)
+          const mod = moduleGraph.idToModuleMap.get(id)
           const rawSourceMap = mod?.ssrTransformResult?.map
 
           if (!rawSourceMap) {
@@ -35,7 +44,8 @@ export function ssrRewriteStacktrace(
 
           const pos = originalPositionFor(traced, {
             line: Number(line) - offset,
-            column: Number(column),
+            // stacktrace's column is 1-indexed, but sourcemap's one is 0-indexed
+            column: Number(column) - 1,
           })
 
           if (!pos.source || pos.line == null || pos.column == null) {
@@ -43,7 +53,9 @@ export function ssrRewriteStacktrace(
           }
 
           const trimmedVarName = varName.trim()
-          const source = `${pos.source}:${pos.line}:${pos.column}`
+          const sourceFile = path.resolve(path.dirname(id), pos.source)
+          // stacktrace's column is 1-indexed, but sourcemap's one is 0-indexed
+          const source = `${sourceFile}:${pos.line}:${pos.column + 1}`
           if (!trimmedVarName || trimmedVarName === 'eval') {
             return `    at ${source}`
           } else {